pyproject.toml
[tool.poetry]
name = "2024-fintech-final-rag-app"
version = "0.1.0"
description = ""
authors = ["r12922064@csie.ntu.edu.tw>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
neo4j = "^5.23.1"
pypdf3 = "^1.0.6"
langchain = "^0.2.12"
python-dotenv = "^1.0.1"
langchain-community = "^0.2.11"
beautifulsoup4 = "^4.12.3"
google-generativeai = "0.7.2"
langchain-google-genai = "1.0.10"
langchainhub = "^0.1.21"
langsmith = "^0.1.137"
pytesseract = "^0.3.13"
pdfplumber = "^0.11.4"
pdf2docx = "^0.5.8"
spire-doc = "^12.7.1"
matplotlib = "^3.9.2"
pyqt6 = "^6.7.1"
pandas = "^2.2.3"
pymupdf4llm = "^0.0.17"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
